refactor(zestymarkdownparser.js): componetize all markdown components
